+++ packages/nx-executors/plugin/index.ts
@@ -1,2 +1,0 @@
-export type { PluginOptions } from '../src/plugins/plugin'
-export { createNodes, createNodesV2 } from '../src/plugins/plugin'