+++ packages/nx-executors/plugin/jest.js
@@ -1,0 +1,6 @@
+const import_jest = require('../dist/plugins/jest')
+
+module.exports = {
+  createNodes: import_jest.createNodes,
+  createNodesV2: import_jest.createNodesV2
+}